requirements.txt
websocket-client
websockets
twisted
autobahn
